pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "apache-airflow-provider-ovhcloud-ai"
dynamic = ["version"]
description = "Apache Airflow Provider for OVHcloud AI"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "Elias TOURNEUX", email = "elias.tourneux.ext@ovhcloud.com"}
]
keywords = ["airflow", "provider", "ovhcloud", "ai", "llm", "openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
]

dependencies = [
    "apache-airflow>=2.3.0",
    "requests>=2.25.0",
    "pytest>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/ovh/apache-airflow-provider-ovhcloud-ai"
Documentation = "https://github.com/ovh/apache-airflow-provider-ovhcloud-ai#readme"
Repository = "https://github.com/ovh/apache-airflow-provider-ovhcloud-ai"
"Bug Tracker" = "https://github.com/ovh/apache-airflow-provider-ovhcloud-ai/issues"

[project.entry-points."apache_airflow_provider"]
provider_info = "apache_airflow_provider_ovhcloud_ai:get_provider_info"
